+++ R/zzz.R
@@ -1,6 +1,17 @@
+# Package Environment
+.errorist_env = new.env()
+.errorist_env$warn_level = 0
+
+# Retrieve settings on load
+setup_errorist_environment = function() {
+  .errorist_env$op = options()
+  .errorist_env$warn_level =  getOption("warn", 0)
+}
+
 .onAttach = function(libname, pkgname) {
 
   autoload_config = getOption("errorist.autoload", TRUE)
+  setup_errorist_environment()
 
   if(!is.logical(autoload_config)) {